pageserver: remove tenant deletion-after-restart code
